refactor: 불필요한 클래스 삭제, Security 관련 클래스들 패키지 이동
