+++ src/main/java/com/saebyeok/saebyeok/security/User.java
@@ -1,16 +1,13 @@
-package com.saebyeok.saebyeok.domain;
+package com.saebyeok.saebyeok.security;
 
+import com.saebyeok.saebyeok.domain.Role;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
-import javax.persistence.*;
-import java.time.LocalDateTime;
 import java.util.Collection;
 import java.util.Collections;
